fix: try alternate user agents (#165)
* Avoid 301 for initial navigation

Navigating to NQ%3DF results in a redirection to
NQ%3DF/. We can avoid this by adjusting the starting url.

* Try multiple user agents

The server end point doesn't seem to like standard user-agents. We
are now trying another simpler agents when 429 is returned.

CrumbCoordinator tracks the successful user-agent which is then used
by the data coordinators.

* Skip agent if already failed