fix: special handling for GBp
The currency GBp is special and is used by LGEN.L . This just bears a
case difference from GBP.

We cannot use standard conversion since the upper cases symbol is the
same. Sibce the conversion factor to GBP is 1/100 so we will special
case this and tack on this additional conversion. The financial_currency
in this case is GBP.